requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-amazon==6.2.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-postgres==5.3.1
apache-airflow-providers-sqlite==3.3.1
pymongo==4.3.3
great-expectations==0.15.31
pandas==1.3.5
numpy==1.21.6
psycopg2==2.9.5
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.11
SQLAlchemy-Utils==0.38.3